Changed behaviour of scroll icon
The icon now rotates based on the view that is currently being shown.

This makes it somewhat useful when the browser's address bar hides, making the icon visible. Still pressing the icon to go up does not reveal the address bar so you can still see the images below.

Not quite sure what to do.